consolidate styling between plantpages in /plant-page route
