+++ app/plant-page/all-plants/[plantId]/style.ts
